requirements.txt
aiohttp==3.9.3
aiopg==1.4.0
aiosignal==1.3.1
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==4.3.0
APScheduler==3.10.4
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
bitarray==2.9.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cytoolz==0.12.3
ecdsa==0.18.0
eth-account==0.11.0
eth-hash==0.7.0
eth-keyfile==0.8.0
eth-keys==0.5.0
eth-rlp==1.0.1
eth-typing==4.0.0
eth-utils==4.0.0
eth_abi==5.0.1
exceptiongroup==1.2.0
fastapi==0.110.0
frozenlist==1.4.1
h11==0.14.0
hexbytes==0.3.1
idna==3.6
iniconfig==2.0.0
iso8601==1.1.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
loguru==0.7.2
lru-dict==1.2.0
multidict==6.0.5
packaging==24.0
parsimonious==0.9.0
pluggy==1.4.0
protobuf==5.26.1
psycopg2-binary==2.9.9
pyasn1==0.6.0
pycryptodome==3.20.0
pydantic==1.10.7
pydantic_core==2.16.3
pypika-tortoise==0.1.6
pytest==8.1.1
python-dotenv==1.0.1
python-jose==3.3.0
pytz==2024.1
pyunormalize==15.1.0
PyYAML==6.0.1
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
retrying==1.3.4
rlp==4.0.0
rpds-py==0.18.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
tomli==2.0.1
toolz==0.12.1
tortoise-orm==0.20.0
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
uvicorn==0.29.0
web3==6.16.0
websockets==12.0
yarl==1.9.4
